Implemented database trigger for score
Actually created custom value resolver for automapper so we don't need to store score and we can just calculate it on sight

+++ BLL/Dtos/OutDto/GameForResponceDto.cs
@@ -7,7 +7,7 @@
         public int DeveloperId { get; set; }
         public string Name { get; set; } = null!;
         public DateTime ReleaseDate { get; set; }
-        public byte? Score { get; set; }
+        public double? Score { get; set; }
         public string? Description { get; set; }
         public decimal Price { get; set; }
     }
